SbsaQemu: get the information of CPU topology via SMC calls
Provide functions to check for CPU topology information:
 - the number of sockets on sbsa-ref platform.
 - the number of clusters in one socket.
 - the number of cores in one cluster.
 - the number of threads in one core.

As SMC calls can return up to 4 return values, the number of sockets,
clusters and cores are read from TF-A using platform specific SMC call.
Number of threads is caluculated using the cpu count and the number of
sockets, clusters and cores.

Signed-off-by: Xiong Yining <xiongyining1480@phytium.com.cn>
Signed-off-by: Marcin Juszkiewicz <marcin.juszkiewicz@linaro.org>